auth_service/pyproject.toml
[tool.poetry]
name = "Auth-service"
version = "0.1.0"
description = ""
authors = ["Kirill Iakimenkov <kayakimenkov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
redis = "^4.4.2"
fastapi = "^0.104"
pydantic = {extras = ["email"], version = "^2.5.2"}
pydantic-settings = "^2.0.0"
pydantic-extra-types = "^2.1.0"
uvicorn = "^0.22"
gunicorn = "^21.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
passlib = "^1.7.4"
phonenumbers = "^8.13.26"
asyncpg = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlalchemy-utils = "^0.41.1"
werkzeug = "^3.0.1"
python-multipart = "^0.0.6"
orjson = "^3.9.10"
opentelemetry-api = "1.17.0"
opentelemetry-sdk = "1.17.0"
opentelemetry-instrumentation-fastapi = "0.38b0"
opentelemetry-exporter-jaeger = "1.17.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[tool.poetry.group.test.dependencies]
faker = "^19.12.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.0"
factory-boy = "^3.3.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
